Better way of grabbing the skins for Bedrock players
The old way grabbed the skin of the Java player with the same username as the Bedrock player, but this is of course not an ideal method.
The new method uses the XUID (Xbox User Identification) to grab the texture ID from the Geyser API directly and then grabs the actual skin file from the Mojang servers.
It then simply cuts out the head part from the skin and badabingbadabo-am you've got a head icon!

Big thanks to https://github.com/crafatar/crafatar/issues/287